requirements.txt
flask
paho-mqtt
psycopg2-binary
redis
geopy
zlapi
python-dotenv